refactor out generate_recv_code_for_packet
